requirements.txt
pyrogram>=2.0.101
tgcrypto>=1.2.5
motor>=3.3.2
python-dotenv>=1.0.0
psutil>=5.9.8
